Initial texture support.
Quick fixed EJTexture to hold pixels instead of textureId which was used for the 2D stuff. Need to fix this later to support 3D as well as 2D by doing a lazy texture creation fro EJTexture when a 2D image operation is requested.

Lesson05 now works and added to examples.