Switch to a more complete(but hacky) typed array shim.
This uses the ability to use any Object as an array. The shims are just regular objects masquerading as typed arrays. This allows us to use WebGL code as is since the new typed array shim actually allows for access of elements after creation while the old one didn't. This will hopefully be not needed once typed array support is properly enabled in JavascriptCore.

So with this change replacing the custom glMatrix library with the original (latest) gl-matrix.js which uses typed arrays.

No apparent perf change.