createImageData() pixel allocation change from malloc() to calloc()
The problem with malloc() was that calling createImageData() repeatedly
caused it to return an ImageData object which contained pixel data that was
set to the previously returned ImageData object.

createImageData() should always return an object filled with transparent black,
calloc() achieves this by automatically filling the allocated space with zeroes.