Change _publishEvent to dispatchEvent
Your _publishEvent function is basically the same thing as the native dispatchEvent method which is a public method that would be nice to call from other places. Hope you don't mind the change. I was working with window resize events and wishing for this change. Sal will have some updates on this that will include some of my js maybe also.