added SIEmbeddedFileSystem to protect JavaScript source
Programs writing programs! During the build phase embed.php uses xxd to
generate C code for every JavaScript file in the project. This code is
used by the SIEmbeddedFileSystem to access the JavaScript source (via
the original path) without exposing it in an unarchived IPA. copy.php
copies everything but the JavaScript files to the new IPA.